refactor(#45): NotificationData data class 패키지 이동
