+++ app/src/main/java/com/ssafy/neegongnaegong/domain/model/studies/NotificationData.kt
@@ -1,0 +1,10 @@
+package com.ssafy.neegongnaegong.domain.model.studies
+
+/**
+ * 알림 데이터 모델
+ */
+data class NotificationData(
+    val id: Long,
+    val title: String,
+    val dateTime: String,
+)
