+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/group/component/detail/section/NotificationsSection.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ssafy.neegongnaegong.R
-import com.ssafy.neegongnaegong.presentation.group.component.detail.NotificationData
+import com.ssafy.neegongnaegong.domain.model.studies.NotificationData
 import com.ssafy.neegongnaegong.presentation.group.component.detail.NotificationWindow
 import com.ssafy.neegongnaegong.presentation.ui.theme.NeeGongNaeGongTheme
